backend/srec/requirements.txt
Pillow>=10.3
cairosvg>=2.7
xxhash>=3.4
sentence-transformers>=2.7
chromadb>=0.5
tqdm>=4.66
